requirements.txt
absl-py==1.4.0
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
boltons==23.0.0
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.4
click==8.1.6
comm==0.1.3
contourpy==1.1.0
cryptography==39.0.1
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dnspython==2.3.0
einops==0.6.1
exceptiongroup==1.1.1
executing==0.8.3
expecttest==0.1.4
fastjsonschema==2.17.1
filelock==3.12.2
fonttools==4.41.1
fqdn==1.5.1
fsspec==2023.6.0
gensim==4.3.1
glob2==0.7
gmpy2==2.1.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
googletrans==3.0.0
grpcio==1.56.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.16.4
hydra-core==1.3.2
hyperframe==5.2.0
hypothesis==6.75.2
identify==2.5.26
idna==2.10
imageio==2.31.1
ipykernel==6.24.0
ipython==8.12.0
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
kiwisolver==1.4.4
lazy_loader==0.3
libarchive-c==2.9
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.7.2
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python-headless==4.8.0.74
overrides==7.3.1
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
pkginfo==1.9.6
platformdirs==3.9.1
pluggy==1.0.0
pre-commit==3.3.3
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat>=0.6.4
pycparser==2.21
Pygments==2.15.1
pynvml==11.5.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-etcd==0.4.5
python-json-logger==2.0.7
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qudida==0.0.4
referencing==0.30.0
regex==2023.6.3
requests==2.29.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
Send2Trash==1.8.2
setuptools==65.6.3
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
stack-data==0.2.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
terminado==0.17.1
textaugment==1.3.4
textblob==0.17.1
threadpoolctl==3.2.0
tifffile==2023.7.18
timm==0.9.2
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchdata==0.6.1
torchelastic==0.2.2
torchtext==0.15.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.7.1
transformers==4.26.1
triton==2.0.0
types-dataclasses==0.6.6
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.15
virtualenv==20.24.1
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
wheel==0.38.4
zstandard==0.19.0